+++ mjml/__init__.py
@@ -6,11 +6,13 @@ from mjml.errors import (
     ValidationLevel,
     ValidationRule,
 )
+from mjml.helpers.includes import IncludePolicy
 from mjml.mjml2html import ParseResult, mjml_to_html, validate
 
 
 __all__ = [
     'Include',
+    'IncludePolicy',
     'MJMLValidationErrors',
     'mjml_to_html',
     'ParseResult',